refactored helper functions, renamed variables, + comments
