+++ static/books/js/books.js
@@ -1,4 +1,4 @@
-import { getCoverImg, getDetails, displayArchiveOptions } from "../../global/js/sections.js"
+import { getCoverImg, getDetails, displayArchiveOptions } from "../../global/js/helpers.js"
 
 // the only fields that will be displayed when searching
 const FIELDS = {
